refactor: rename secure storage subpath to expo-secure-store
