+++ package.json
@@ -105,10 +105,10 @@
     "viem": {
       "optional": true
     },
-    "expo-secure-store": {
+    "expo-web-browser": {
       "optional": true
     },
-    "expo-web-browser": {
+    "expo-secure-store": {
       "optional": true
     },
     "@react-native-async-storage/async-storage": {
@@ -134,8 +134,8 @@
   "exports": {
     ".": {
       "types": "./dist/index.d.ts",
-      "react-native": "./dist/index.js",
       "src": "./src/index.ts",
+      "react-native": "./dist/index.js",
       "default": "./dist/index.js"
     },
     "./server": {
@@ -165,39 +165,39 @@
     },
     "./react-native": {
       "types": "./dist/react-native/index.d.ts",
-      "react-native": "./dist/react-native/index.js",
       "src": "./src/react-native/index.ts",
+      "react-native": "./dist/react-native/index.js",
       "default": "./dist/react-native/index.js"
     },
     "./mobileWebAuth": {
       "types": "./dist/mobileWebAuth/index.d.ts",
-      "react-native": "./dist/mobileWebAuth/index.js",
       "src": "./src/mobileWebAuth/index.ts",
+      "react-native": "./dist/mobileWebAuth/index.js",
       "default": "./dist/mobileWebAuth/index.js"
     },
     "./mobileWebAuth/tempoWallet": {
       "types": "./dist/mobileWebAuth/tempoWallet.d.ts",
-      "react-native": "./dist/mobileWebAuth/tempoWallet.js",
       "src": "./src/mobileWebAuth/tempoWallet.ts",
+      "react-native": "./dist/mobileWebAuth/tempoWallet.js",
       "default": "./dist/mobileWebAuth/tempoWallet.js"
     },
     "./react-native/async-storage": {
       "types": "./dist/react-native/asyncStorage.d.ts",
-      "react-native": "./dist/react-native/asyncStorage.js",
       "src": "./src/react-native/asyncStorage.ts",
+      "react-native": "./dist/react-native/asyncStorage.js",
       "default": "./dist/react-native/asyncStorage.js"
     },
     "./react-native/expo-web-browser": {
       "types": "./dist/react-native/expoWebBrowser.d.ts",
-      "react-native": "./dist/react-native/expoWebBrowser.js",
       "src": "./src/react-native/expoWebBrowser.ts",
+      "react-native": "./dist/react-native/expoWebBrowser.js",
       "default": "./dist/react-native/expoWebBrowser.js"
     },
-    "./react-native/secure-storage": {
-      "types": "./dist/react-native/secureStorage.d.ts",
-      "react-native": "./dist/react-native/secureStorage.js",
-      "src": "./src/react-native/secureStorage.ts",
-      "default": "./dist/react-native/secureStorage.js"
+    "./react-native/expo-secure-store": {
+      "types": "./dist/react-native/expoSecureStore.d.ts",
+      "src": "./src/react-native/expoSecureStore.ts",
+      "react-native": "./dist/react-native/expoSecureStore.js",
+      "default": "./dist/react-native/expoSecureStore.js"
     }
   }
 }
