+++ src/react-native/index.ts
@@ -1,2 +1,2 @@
 export { asyncStorage } from './asyncStorage.js'
-export { secureStorage } from './secureStorage.js'
+export { secureStorage } from './expoSecureStore.js'
